feat(PA4): support context switch

+++ nanos-lite/include/fs.h
@@ -13,6 +13,6 @@ size_t fs_write(int fd, const void *buf, size_t len);
 size_t fs_lseek(int fd, size_t offset, int whence);
 int fs_close(int fd);
 char* get_filename(int fd) ;
-size_t sys_write(intptr_t fd, void* buf, size_t len);
+size_t sys_write(int fd, void* buf, size_t len);
 size_t sys_read(int fd, void* buf, size_t len);
 #endif
